+++ crates/b3/src/likelihood/gamma.rs
@@ -53,7 +53,7 @@ impl GammaLikelihood {
 		let mut calculators = Vec::new();
 		let mut transitions = Vec::new();
 
-		let num_nodes = tree.get().num_nodes();
+		let num_nodes = tree.get().num_nodes() as usize;
 
 		for _ in 0..num_categories {
 			calculators.push(calculator
